+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityConditionModelFactory.java
@@ -26,6 +26,8 @@ import is.codion.framework.model.EntityConditionModelFactory;
 import is.codion.swing.common.model.component.combobox.FilterComboBoxModel;
 import is.codion.swing.framework.model.component.EntityComboBoxModel;
 
+import java.util.Optional;
+
 import static is.codion.swing.framework.model.SwingForeignKeyConditionModel.swingForeignKeyConditionModel;
 
 /**
@@ -42,15 +44,12 @@ public class SwingEntityConditionModelFactory extends EntityConditionModelFactor
 	}
 
 	@Override
-	public ColumnConditionModel<? extends Attribute<?>, ?> createConditionModel(Attribute<?> attribute) {
-		if (!includes(attribute)) {
-			throw new IllegalArgumentException("Condition model for attribute: " + attribute + " is not included");
-		}
+	public Optional<ColumnConditionModel<? extends Attribute<?>, ?>> createConditionModel(Attribute<?> attribute) {
 		if (attribute instanceof ForeignKey) {
 			ForeignKey foreignKey = (ForeignKey) attribute;
 			if (definition(foreignKey.referencedType()).smallDataset()) {
-				return swingForeignKeyConditionModel(foreignKey,
-								this::createEqualComboBoxModel, this::createInSearchModel);
+				return Optional.of(swingForeignKeyConditionModel(foreignKey,
+								this::createEqualComboBoxModel, this::createInSearchModel));
 			}
 		}